influx_nlp/pyproject.toml
[project]
name = "influx-nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "deep-translator>=1.11.4",
    "flask>=3.1.0",
    "ginza>=5.2.0",
    "inline-snapshot>=0.23.2",
    "ipykernel>=6.29.5",
    "ja-ginza>=5.2.0",
    "spacy>=3.8.7",
    "stanza>=1.10.1",
    "sudachipy>=0.6.10",
    "sudachidict-full>=20250515",
    "wanakana-python>=1.1.4",
]

[dependency-groups]
dev = [
    "black>=24.4.2",
    "fugashi[unidic]>=1.5.1",
    "nltk>=3.9.1",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "pykakasi>=2.3.0",
    "spacy>=3.8.7",
    "tabulate>=0.9.0",
    "unidic>=1.1.0",
]

[tool.black]
line-length = 120
target-version = ['py312']
